refactor: Lines 도메인 생성, 책임 분리
도메인의 역할에 맞게 패키지 분리
