+++ docs/README.md
@@ -163,8 +163,8 @@
 - [X] Fare 의 상수를 `50km, 10km` 가 아닌 `장거리, 단거리` 와 같은 이름으로 수정
 - [X] Name 의 Pattern 은 비싼 자원이므로 재사용 하기
 - [ ] 알고리즘 or 라이브러리와 같이 변경이 될 수 있는 부분은 인터페이스를 사용
-- [ ] Lines 도메인 만들기
-- [ ] PathCalculator 자료형이 포함된 변수명 수정
+- [X] Lines 도메인 만들기
+- [X] PathCalculator 자료형이 포함된 변수명 수정
 - [ ] 에러 메세지를 더 친절하게 작성 (LineService)
 - [ ] 900 으로 하드코딩 된 부분 수정
 - [ ] PathService 의 DB O(n)번 접근하는 비용 문제 해결
