+++ src/test/java/wooteco/subway/domain/ColorTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.line.Color;
+
 class ColorTest {
     @Test
     @DisplayName("색상이 null이면 예외를 반환한다.")
