+++ src/test/java/wooteco/subway/domain/LineTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.line.Line;
+
 class LineTest {
 
     @Test
