+++ src/test/java/wooteco/subway/domain/PathCalculatorTest.java
@@ -7,6 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.path.PathCalculator;
+
 public class PathCalculatorTest {
 
     @Test
